requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.1
marshmallow==3.20.1
PyMySQL==1.1.0
python-dotenv==1.0.0
Werkzeug==3.0.1
WTForms==3.1.1
email-validator==2.1.0
